requirements.txt
Django>=5.0,<6.0
mysqlclient>=2.2
Pillow>=10.0
celery>=5.4
redis>=5.0
python-dotenv>=1.0
pytest
pytest-django
gunicorn
factory_boy

